(follow-up) Fix insert circ rules since bug 18925
I messed everything up in the last commit.
I cherry-picked a local branch that contained changes for the whole circ
rules move. Here (18925) we only moved maxissueqty and maxonsiteissueqty

TODO: Things are going to be even more dirty when other changes will be
pushed, we will need to rework this code to make it more robust.