Issue #36 - Add the ability to launch a custom SQL file
Devs have different needs, it would be great to have the ability to give a .sql
file to execute at the end of the installation process.

Note: we are going to assume that everybody wants a "shared" directory in their koha root directory.
We could make it configurable later.
Also this directory could be reused for other uses: dotfiles or history files
shared with the docker container.

Test plan:
0/ Apply this patch
1/ % reset_all
2/ You will notice "There is no custom.sql ($shared_dir/custom.sql) file, skipping."
3/ % mkdir /kohadev/koha/shared
4/ % echo 'shared/' >> /kohadev/koha/.git/info/exclude
5/ % cp custom-sample.sql /kohadev/koha/shared
6/ Edit custom-sample.sql and uncomment everything
7/ % reset_all
=> The new prefs' values have been used