Issue #29 - Support for UNIMARC
What is done here:
* update the records data for UNIMARC, they pass the inconsistency check
script
* configure lang="en"
fr-FR sample data set to fr-FR, but we do not regenerate the templates
* populate_db.pl retrieve the UNIMARC sample data from fr-FR
* Branchcode=CPL and categorycode=S are the default values if they
exist. Otherwise the first from the tables (branches and categories) is
retrieved.

Known limits:
Will only work after 18.11

Test plan:
Apply Koha patches from bug 24851 [1] and bug 24856 [2]
Go to koha-testing-docker, edit .env and set
 DEBUG_GIT_REPO_MISC4DEV_URL=https://gitlab.com/joubu/koha-misc4dev.git
 DEBUG_GIT_REPO_MISC4DEV_BRANCH=issue_29
 KOHA_MARC_FLAVOUR=unimarc
Note that the 2 DEBUG vars are only needed because this patch is not into
koha-misc4dev:master yet
Start the docker container
Wait for everything to be ready, then start a search
=> Confirm that results are returned
Edit a record, save, play with everything you want
This is a first step, things may need to be fixed (later)

[1] https://bugs.koha-community.org/bugzilla3/show_bug.cgi?id=24851
[2] https://bugs.koha-community.org/bugzilla3/show_bug.cgi?id=24856